fix(realm-compat): split machine oracle tests and close fuel bounds
Move the crate-private machine oracle into a child tests module so every source file remains within the repository line limit. Keep the existing safety comments and fixture behavior intact, and prove exact maximum fuel acceptance plus max-plus-one and u64::MAX rejection while retaining zero-fuel rejection.

Tracking #1564

AI / Tool Assistance: OpenAI Codex implemented the exact-head AMEND for the Realm compatibility machine oracle on the isolated branch.
